// ui-dock/src/app/models/headerMenu.model.ts
/** *****************************************************************************
Licensed Materials - Property of IBM 6949-XXX ã Copyright IBM Corp. 2019
All Rights Reserved US Government Users Restricted Rights - Use, duplication 
or disclosure restricted by GSA ADP Schedule Contract with IBM Corp.
***************************************************************************** */

export class HeaderMenuModel {
    name: string;
    code: string;
    seq: number;
    isSelected: boolean;
    isVisible: boolean;
    isClickable: boolean;
    route: string;
}